generics/mixins: Move GenericAPIView code to GenericModelMixin
There are third party libraries, for example ADRF, which provide
their own customized implementation of `APIView`.

It is useful to be able to create `GenericAPIView` variants based
on these customized `APIView` variants.

However, as `GenericAPIView` inherits directly from Django Rest
Framework's `APIView`, there is no clean way to achieve this without
duplicating the code of `GenericAPIView`.

Resolve this by moving the code of `GenericAPIView` to a mixin
called `GenericModelMixin` and creating `GenericAPIView` by mixing
`GenericModelMixin` with Django Rest Framework's `APIView`.

This way, other libraries like ADRF can mix `GenericModelMixin` with
their customized `APIView` variant and resuse the code of the
`GenericModelMixin` without duplicating it.
